PE-1769: unit test refactoring.
