+++ ddf-test/src/main/scala/io/ddf/test/it/etl/JoinHandlerBaseSuite.scala
@@ -18,10 +18,12 @@
  */
 package io.ddf.test.it.etl
 
+import java.util
 import java.util.Collections
 
 import io.ddf.DDF
 import io.ddf.etl.Types.JoinType
+import io.ddf.spark.etl.JoinHandler
 import io.ddf.test.it.BaseSuite
 import org.scalatest.Matchers
 
@@ -89,4 +91,14 @@ trait JoinHandlerBaseSuite extends BaseSuite with Matchers {
     joinedDDF.getNumRows should be(33)
   }
 
+  test("inner join suffix") {
+    val left_ddf = loadMtCarsDDF()
+    val right_ddf = loadCarOwnersDDF()
+    val ddf = left_ddf.join(right_ddf, JoinType.INNER, Collections.singletonList("cyl"),null,null, "_left", "_right")
+    val colNames = ddf.getSchema.getColumnNames
+    colNames.contains("cyl_left") should be(true)
+    colNames.contains("cyl_right") should be(true)
+    colNames.contains("disp_left") should be(true)
+    colNames.contains("disp_right") should be(true)
+  }
 }
